Release 0.15.9.1
Improvement  when a peer disconnecting is detected a resend certificate
validation is triggered.

Improvement to ensure that if in case something else happened in the
bridge process validation of a certificate and no answer comes after a
certain time of no answer the certificate validation is triggered.

Improvement is implemented to avoid blocks being received with empty
certificates, if it was the case then we mark the block as not received
to force the blockchain to sync it again.